add new branch rm package stuff mv python scripts back to home project root delete package building readme update readme
