--- berrycam_0.1-1_armhf/DEBIAN/postinst
@@ -1,4 +1,0 @@
-#!/bin/bash
-sudo systemctl enable /usr/bin/berryCam/berryCam.service
-sudo systemctl daemon-reload
-sudo systemctl start berryCam
