--- berrycam_0.1-1_armhf/DEBIAN/prerm
@@ -1,4 +1,0 @@
-#!/bin/bash
-sudo systemctl stop berryCam
-sudo systemctl disable berryCam
-sudo systemctl daemon-reload
